major update on facility client relation

+++ web-service/src/main/java/com/example/webservice/service/FacilityService.java
@@ -2,13 +2,17 @@ package com.example.webservice.service;
 
 import com.example.webservice.model.Client;
 import com.example.webservice.model.Facility;
+import com.example.webservice.model.Listing;
 import com.example.webservice.model.model.FacilityRequestDTO;
 import com.example.webservice.repository.ClientRepository;
 import com.example.webservice.repository.FacilityRepository;
+import com.example.webservice.repository.ListingRepository;
 import org.apache.velocity.exception.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
@@ -17,6 +21,7 @@ public class FacilityService {
 
     private final ClientRepository clientRepository;
     private final FacilityRepository facilityRepository;
+    private final ListingRepository listingRepository;
 
     /**
      * Constructs a new FacilityService with the specified repositories.
@@ -26,19 +31,21 @@ public class FacilityService {
      */
     @Autowired
     public FacilityService(ClientRepository clientRepository,
-                           FacilityRepository facilityRepository) {
+                           FacilityRepository facilityRepository,
+                           ListingRepository listingRepository) {
         this.clientRepository = clientRepository;
         this.facilityRepository = facilityRepository;
+        this.listingRepository = listingRepository;
     }
 
     /**
-     * Retrieves a list of all facilities.
+     * Retrieves a list of all public facilities.
      *
-     * @return the list of all facilities
+     * @return the list of all public facilities
      */
     @Transactional
     public List<Facility> getAllFacilities() {
-        return facilityRepository.findAll();
+        return facilityRepository.findPublicFacilities();
     }
 
     /**
@@ -54,6 +61,35 @@ public class FacilityService {
                 .orElseThrow(() -> new ResourceNotFoundException("Facility not found with id: " + id));
     }
 
+    /**
+     * Creates a facility with the given data. Requires authentication.
+     *
+     * @param clientId         the ID of the client making the update request
+     * @param facility  the data transfer object containing new facility details
+     * @return the updated facility
+     * @throws ResourceNotFoundException if no client or facility is found with the given IDs or if authentication fails
+     */
+    @Transactional
+    public Facility createFacility(Long clientId,
+                                   FacilityRequestDTO facility) {
+        Client c = clientRepository.findById(clientId)
+                .orElseThrow(() -> new ResourceNotFoundException("Client not found with id: " + clientId));
+
+        try {
+            Facility newFacility = new Facility();
+            newFacility.setAssociated_distributorID(c.getClientID());
+            newFacility.setLatitude(facility.getLatitude());
+            newFacility.setLongitude(facility.getLongitude());
+            newFacility.setPublic(facility.getPublic());
+            newFacility.setEmail(facility.getEmail());
+            newFacility.setPhone(facility.getPhone());
+            newFacility.setHours(facility.getHours());
+            return facilityRepository.save(newFacility);
+        } catch (Exception e) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST,
+                    e.getMessage());
+        }
+    }
 
     /**
      * Updates a facility with the given data. Requires authentication.
@@ -76,13 +112,18 @@ public class FacilityService {
         if (!auth.equals(c.getAuthentication())) {
             throw (new ResourceNotFoundException("wrong auth"));
         }
-        if (!id.equals(c.getAssociatedFacility().getFacilityID())) {
+
+        Facility f = facilityRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Facility not found with id: " + id));
+
+        if (!clientId.equals(f.getAssociated_distributorID())) {
             throw (new ResourceNotFoundException("wrong facilityID"));
         }
 
         return facilityRepository.findById(id)
                 .map(facility -> {
                     facility.setFacilityID(id);
+                    facility.setAssociated_distributorID(clientId);
                     facility.setLatitude(updatedFacility.getLatitude());
                     facility.setLongitude(updatedFacility.getLongitude());
                     facility.setPublic(updatedFacility.getPublic());
@@ -97,11 +138,24 @@ public class FacilityService {
     /**
      * Deletes a facility based on its ID.
      *
+     * @param clientId the ID of the client associated with the facility
      * @param id the ID of the facility to delete
      * @return true if the facility was found and deleted, false otherwise
      */
-    public boolean deleteFacility(Long id) {
+    public boolean deleteFacility(Long clientId, Long id) {
+
+
         if (facilityRepository.existsById(id)) {
+            Facility f = facilityRepository.findById(id)
+                    .orElseThrow(() -> new ResourceNotFoundException("Facility not found with id: " + id));
+
+            if (!clientId.equals(f.getAssociated_distributorID())) {
+                throw (new ResourceNotFoundException("unmatched facilityID and client"));
+            }
+            List<Listing> listingID = listingRepository.findListingsByFacilityID(id);
+            for (Listing l : listingID) {
+                listingRepository.deleteById(l.getListingID());
+            }
             facilityRepository.deleteById(id);
             return true;
         } else {
